+++ v2/cmd/gazctl/journals_reset_head.go
@@ -12,6 +12,44 @@ type cmdJournalResetHead struct {
 	Selector string `long:"selector" short:"l" required:"true" description:"Label Selector query to filter on"`
 }
 
+func init() {
+	_ = mustAddCmd(cmdJournals, "reset-head", "Reset journal append offset (disaster recovery)", `
+Reset the append offset of journals.
+
+Gazette appends are transactional: all brokers must agree on the exact offsets
+at which an append operation will be written into a journal. The offset is an
+explicit participate in the broker's transaction protocol. New participants are
+"caught up" on the current offset by participating in broker transactions, and
+brokers will delay releasing responsibility for a journal until all peers have
+participated in a synchronizing transaction. This makes Gazette tolerant to up
+to R-1 independent broker process failures, where R is the replication factor
+of the journal.
+
+However, disasters and human errors do happen, and if R or more independent
+failures occur, Gazette employs a fail-safe to minimize the potential for a
+journal offset to be written more than once: brokers require that the remote
+fragment index not include a fragment offset larger than the append offset known
+to replicating broker peers, and will refuse the append if this constraint is
+violated.
+
+Eg, If N >= R failures occur, then the set of broker peers of a journal will not
+have participated in an append transaction; their append offset will be zero,
+which is less than the maximum offset contained in the fragment store. The
+brokers will refuse all appends to preclude double-writing of an offset.
+
+This condition must be explicitly cleared by the Gazette operator using the
+reset-head command. The operator should delay running reset-head until absolutely
+confident that all journal fragments have been persisted to cloud storage (eg,
+because all previous broker processes have exited).
+
+Then, the effect of reset-head is to jump the append offset forward to the
+maximum indexed offset, allowing new append operations to proceed.
+
+reset-head is safe to run against journals which are in a fully consistent state,
+though it is likely to fail harmlessly if the journal is being actively written.
+`, &cmdJournalResetHead{})
+}
+
 func (cmd *cmdJournalResetHead) Execute([]string) error {
 	startup()
